第十五次上传： 1. 使用QFile，解决了文件读取的路径问题 2. 给Tower添加了lvlup()方法，通过鼠标事件监控。现在可以给防御塔升级。 3. 在Chapter中添加了showTowerRange()方法， 并通过鼠标事件监控。现在将鼠标移至防御塔的位置可显示攻击范围。

+++ Tower_Defence/Tower.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 using namespace std;
 
-int Tower::max_range = TOWER_RANGE;
+double Tower::max_range = TOWER_RANGE;
 int Tower::price = TOWER_PRICE;
 
 Tower::Tower(string type) : Object(type)
@@ -16,7 +16,6 @@ Tower::Tower(string type) : Object(type)
 void Tower::set_target(Monster * t)
 {
     this->_target = t;
-   // std::cout<<t<<endl;
     if (t != NULL) this->_start = clock();
 }
